refactor: auth 패키지 global 패키지로 이동 #131
