--- src/main/java/com/bookripple/api/domain/auth/entity/RefreshToken.java
@@ -1,33 +1,0 @@
-package com.bookripple.api.domain.auth.entity;
-
-import com.bookripple.api.global.entity.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "refresh_token")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Getter
-@Builder
-public class RefreshToken extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private Long memberId;
-
-    @Column(nullable = false, unique = true, columnDefinition = "TEXT")
-    private String token;
-
-    @Column(nullable = false)
-    private LocalDateTime expiresAt;
-
-    public boolean isExpired() {
-        return LocalDateTime.now().isAfter(expiresAt);
-    }
-}
