--- src/main/java/com/bookripple/api/domain/auth/repository/RefreshTokenRepository.java
@@ -1,19 +1,0 @@
-package com.bookripple.api.domain.auth.repository;
-
-import com.bookripple.api.domain.auth.entity.RefreshToken;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-
-    Optional<RefreshToken> findByToken(String token);
-
-    Optional<RefreshToken> findByMemberId(Long memberId);
-
-    void deleteByMemberId(Long memberId);
-
-    void deleteByToken(String token);
-}
