+++ src/main/java/com/bookripple/api/global/auth/dto/AuthResDto.java
@@ -1,4 +1,4 @@
-package com.bookripple.api.domain.auth.dto;
+package com.bookripple.api.global.auth.dto;
 
 import lombok.Builder;
 import lombok.Getter;
