+++ src/main/java/com/bookripple/api/global/auth/entity/RefreshToken.java
@@ -1,0 +1,41 @@
+package com.bookripple.api.global.auth.entity;
+
+import com.bookripple.api.global.entity.BaseEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import java.time.LocalDateTime;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "refresh_token")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Getter
+@Builder
+public class RefreshToken extends BaseEntity {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @Column(nullable = false)
+  private Long memberId;
+
+  @Column(nullable = false, unique = true, columnDefinition = "TEXT")
+  private String token;
+
+  @Column(nullable = false)
+  private LocalDateTime expiresAt;
+
+  public boolean isExpired() {
+    return LocalDateTime.now().isAfter(expiresAt);
+  }
+}
