+++ src/main/java/com/bookripple/api/global/auth/service/AuthService.java
@@ -1,22 +1,22 @@
-package com.bookripple.api.domain.auth.service;
+package com.bookripple.api.global.auth.service;
 
-import com.bookripple.api.common.code.AuthErrorCode;
-import com.bookripple.api.common.code.CommonErrorCode;
-import com.bookripple.api.common.code.MemberErrorCode;
-import com.bookripple.api.common.error.ApiException;
-import com.bookripple.api.domain.auth.dto.AuthReqDto;
-import com.bookripple.api.domain.auth.dto.AuthResDto;
-import com.bookripple.api.domain.auth.entity.RefreshToken;
-import com.bookripple.api.domain.auth.repository.RefreshTokenRepository;
+import com.bookripple.api.domain.member.code.MemberErrorCode;
 import com.bookripple.api.domain.member.entity.Member;
 import com.bookripple.api.domain.member.enums.LoginType;
 import com.bookripple.api.domain.member.enums.MemberRole;
 import com.bookripple.api.domain.member.repository.MemberRepository;
-import com.bookripple.api.domain.verification.email.service.EmailVerificationService;
 import com.bookripple.api.domain.verification.email.enums.EmailVerificationPurpose;
-import com.bookripple.api.global.service.TokenBlacklistService;
+import com.bookripple.api.domain.verification.email.service.EmailVerificationService;
+import com.bookripple.api.global.auth.code.AuthErrorCode;
+import com.bookripple.api.global.auth.dto.AuthReqDto;
+import com.bookripple.api.global.auth.dto.AuthResDto;
+import com.bookripple.api.global.auth.entity.RefreshToken;
+import com.bookripple.api.global.auth.repository.RefreshTokenRepository;
+import com.bookripple.api.global.code.CommonErrorCode;
 import com.bookripple.api.global.dto.GlobalDto;
+import com.bookripple.api.global.error.ApiException;
 import com.bookripple.api.global.security.JwtTokenProvider;
+import com.bookripple.api.global.service.TokenBlacklistService;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.util.Map;
